pacman: fix for groups in multiple repos From: Wael Nasreddine <wael@siemens-mobiles.org> Subject: [pacman-dev] Problem with pacman 3 and groups
Hi,

There's a problem with pacman well at least the one from VMiklos
darcs, I didn't test the one in arch's repository though but I'm
pretty sure it has the same issue.

If packages in 2 repositories are members of the same group, pacman
sees 2 groups with the same name instead of one group containing
packages from both repo, First i thought it ain't a bug but more or
less future feature but when I tried to install a group, pacman tried
to install packages from the first repository in pacman.conf, and not
all packages.